requirements.txt
matplotlib==3.3.4
numpy==1.19.2
pillow==8.2.0
imagehash==4.2.0